Remove InputLoopService from delayed-delivery Core_9 sample and move logic back to Program class
